# apps/web/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "발주 예측 API",
    "description": "B2G pre-procurement signal intelligence.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/admin/sources/{key}/run": {
      "post": {
        "tags": [
          "admin"
        ],
        "summary": "Run Source",
        "operationId": "run_source_api_admin_sources__key__run_post",
        "parameters": [
          {
            "name": "key",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SourceRunIn"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Run Source Api Admin Sources  Key  Run Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/sources/{key}": {
      "patch": {
        "tags": [
          "admin"
        ],
        "summary": "Patch Source",
        "operationId": "patch_source_api_admin_sources__key__patch",
        "parameters": [
          {
            "name": "key",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SourcePatchIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Patch Source Api Admin Sources  Key  Patch"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/jobs": {
      "get": {
        "tags": [
          "admin"
        ],
        "summary": "Jobs",
        "operationId": "jobs_api_admin_jobs_get",
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "name": "job",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Job"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "default": 100,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/JobRunOut"
                  },
                  "title": "Response Jobs Api Admin Jobs Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/jobs/{run_id}/retry": {
      "post": {
        "tags": [
          "admin"
        ],
        "summary": "Retry Job",
        "operationId": "retry_job_api_admin_jobs__run_id__retry_post",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Run Id"
            }
          }
        ],
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Retry Job Api Admin Jobs  Run Id  Retry Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/review": {
      "get": {
        "tags": [
          "admin"
        ],
        "summary": "Review Queue",
        "operationId": "review_queue_api_admin_review_get",
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "open",
              "title": "Status"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "default": 50,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ReviewItemOut"
                  },
                  "title": "Response Review Queue Api Admin Review Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/review/{item_id}": {
      "post": {
        "tags": [
          "admin"
        ],
        "summary": "Decide Review",
        "operationId": "decide_review_api_admin_review__item_id__post",
        "parameters": [
          {
            "name": "item_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Item Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReviewDecisionIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReviewItemOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/llm/usage": {
      "get": {
        "tags": [
          "admin"
        ],
        "summary": "Llm Usage",
        "operationId": "llm_usage_api_admin_llm_usage_get",
        "parameters": [
          {
            "name": "days",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 90,
              "minimum": 1,
              "default": 14,
              "title": "Days"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LLMUsageOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/evals": {
      "get": {
        "tags": [
          "admin"
        ],
        "summary": "Evals",
        "operationId": "evals_api_admin_evals_get",
        "parameters": [
          {
            "name": "kind",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Kind"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/EvalRunOut"
                  },
                  "title": "Response Evals Api Admin Evals Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/documents/{document_id}": {
      "get": {
        "tags": [
          "admin"
        ],
        "summary": "Document",
        "operationId": "document_api_admin_documents__document_id__get",
        "parameters": [
          {
            "name": "document_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Document Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentDetailOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Category": {
        "type": "string",
        "enum": [
          "smart_city",
          "safety_cctv",
          "public_sw",
          "ai_data",
          "mobility",
          "energy_env",
          "welfare_care",
          "education",
          "tourism_culture",
          "facility",
          "other"
        ],
        "title": "Category"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "JobRunOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "job": {
            "type": "string",
            "title": "Job"
          },
          "job_id": {
            "type": "string",
            "title": "Job Id"
          },
          "attempt": {
            "type": "integer",
            "title": "Attempt"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "started_at": {
            "type": "string",
            "format": "date-time",
            "title": "Started At"
          },
          "finished_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finished At"
          },
          "duration_ms": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Duration Ms"
          },
          "result": {
            "additionalProperties": true,
            "type": "object",
            "title": "Result"
          }
        },
        "type": "object",
        "required": [
          "id",
          "job",
          "job_id",
          "attempt",
          "status",
          "error",
          "started_at",
          "finished_at",
          "duration_ms",
          "result"
        ],
        "title": "JobRunOut"
      },
      "LLMUsageOut": {
        "properties": {
          "rows": {
            "items": {
              "$ref": "#/components/schemas/LLMUsageRow"
            },
            "type": "array",
            "title": "Rows"
          },
          "spent_today_usd": {
            "type": "number",
            "title": "Spent Today Usd"
          },
          "daily_budget_usd": {
            "type": "number",
            "title": "Daily Budget Usd"
          },
          "cache_hit_rate": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cache Hit Rate"
          },
          "degraded_rate": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Degraded Rate"
          },
          "extractor_mode": {
            "type": "string",
            "title": "Extractor Mode"
          }
        },
        "type": "object",
        "required": [
          "rows",
          "spent_today_usd",
          "daily_budget_usd",
          "cache_hit_rate",
          "degraded_rate",
          "extractor_mode"
        ],
        "title": "LLMUsageOut"
      },
...
      "ReviewDecisionIn": {
        "properties": {
          "action": {
            "type": "string",
            "enum": [
              "approve",
              "reject",
              "edit"
            ],
            "title": "Action"
          },
          "title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          },
          "budget_krw": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Budget Krw"
          },
          "expected_year": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expected Year"
          },
          "commitment": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "committed",
                  "planned",
                  "reviewing",
                  "declined"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Commitment"
          },
          "institution_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Institution Code"
          },
          "category": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Category"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "action"
        ],
        "title": "ReviewDecisionIn"
      },
      "ReviewItemOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "reasons": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Reasons"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "signal": {
            "$ref": "#/components/schemas/SignalOut"
          },
          "institution_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Institution Name"
          }
        },
        "type": "object",
        "required": [
          "id",
          "status",
          "reasons",
          "created_at",
          "signal",
          "institution_name"
        ],
        "title": "ReviewItemOut"
      },
...
      "SourcePatchIn": {
        "properties": {
          "enabled": {
            "type": "boolean",
            "title": "Enabled"
          }
        },
        "type": "object",
        "required": [
          "enabled"
        ],
        "title": "SourcePatchIn"
      },
      "SourceRunIn": {
        "properties": {
          "since": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Since"
          },
          "until": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Until"
          }
        },
        "type": "object",
        "title": "SourceRunIn"
      },
...
    }
  }
}